Refactor dataset info panels
